requirements.txt
transformers
pillow
requests
torch
torchvision
flask
sentencepiece
soundfile
librosa
timm
easyocr